# public/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Digital Bible Platform",
        "description": "Fast, easy, and free API access to video, audio, and text Bibles.",
        "termsOfService": "http://digitalbibleplatform/terms/",
        "contact": {
            "email": "support@digitalbibleplatform.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "4.0.0-beta"
    },
    "servers": [
        {
            "url": "http://localhost:8080/api",
            "description": "Production Server",
            "variables": {
                "schema": {
                    "enum": [
                        "https"
                    ],
                    "default": "https"
                }
            }
        }
    ],
    "paths": {
        "/timestamps": {
            "get": {
                "tags": [
                    "Audio Timing"
                ],
                "summary": "Returns Bible Filesets which have audio timestamps",
                "description": "This call returns a list of fileset that have timestamp metadata associated with them. This data could be used to search audio bibles for a specific term, make karaoke verse & audio readings, or to jump to a specific location in an audio file.",
                "operationId": "v4_timestamps",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/version_number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v4_bible_timestamps"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/bibles": {
            "get": {
                "tags": [
                    "Bibles"
                ],
                "summary": "Returns Bibles based on filter criteria",
                "description": "The base bible route returning by default bibles and filesets that your key has access to",
                "operationId": "v4_bible.all",
                "parameters": [
                    {
                        "name": "language_code",
                        "in": "query",
                        "description": "The iso code to filter results by. This will return results only in the language specified. For a complete list see the `iso` field in the `/languages` route",
                        "schema": {
                            "$ref": "#/components/schemas/Language/properties/iso"
                        }
                    },
                    {
                        "name": "media",
                        "in": "query",
                        "description": "Will filter bibles based upon the media type of their filesets",
                        "schema": {
                            "$ref": "#/components/schemas/BibleFilesetType/properties/set_type_code"
                        },
                        "example": "audio_drama"
                    },
                    {
                        "name": "media_exclude",
                        "in": "query",
                        "description": "Will exclude bibles based upon the media type of their filesets",
                        "schema": {
                            "$ref": "#/components/schemas/BibleFilesetType/properties/set_type_code"
                        },
                        "example": "audio"
                    },
                    {
                        "name": "country",
                        "in": "query",
                        "description": "The iso code to filter results by. This will return results only in the language specified. For a complete list see the `iso` field in the `/country` route",
                        "schema": {
                            "$ref": "#/components/schemas/Country/properties/id"
                        },
                        "example": "21"
                    },
                    {
                        "name": "audio_timing",
                        "in": "query",
                        "description": "This will return results only which have audio timing information available for that bible. The timing information is stored in table bible_file_timestamps.",
                        "schema": {
                            "type": "boolean",
                            "default": false
                        },
                        "example": "true"
                    },
                    {
                        "$ref": "#/components/parameters/page"
                    },
                    {
                        "$ref": "#/components/parameters/limit"
                    },
                    {
                        "$ref": "#/components/parameters/version_number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v4_bible.all"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/bibles/{id}": {
            "get": {
                "tags": [
                    "Bibles"
                ],
                "summary": "Returns detailed metadata for a single Bible",
                "description": "Detailed information for a single Bible",
                "operationId": "v4_bible.one",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Bible/properties/id"
                        }
                    },
                    {
                        "name": "include_font",
                        "in": "query"
                    },
                    {
                        "$ref": "#/components/parameters/version_number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v4_bible.one"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/bibles/{fileset_id}/{book}/{chapter}": {
            "get": {
                "tags": [
                    "Library Text"
                ],
                "summary": "Returns Signed URLs or Text",
                "description": "V4's base fileset route",
                "operationId": "v4_bible.verseinfo",
                "parameters": [
                    {
                        "name": "fileset_id",
                        "in": "query",
                        "description": "The Bible fileset ID",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BibleFileset/properties/id"
                        }
                    },
                    {
                        "name": "book",
                        "in": "query",
                        "description": "The Book ID. For a complete list see the `book_id` field in the `/bibles/books` route.",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Book/properties/id"
                        }
                    },
                    {
                        "name": "chapter",
                        "in": "query",
                        "description": "The chapter number",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/BibleFile/properties/chapter_start"
                        }
                    },
                    {
                        "$ref": "#/components/parameters/version_number"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/v2_text_verse"
                                }
                            },
                            "application/xml": {
                                "schema": {
                                    "$ref": "#/components/schemas/v2_text_verse"
                                }
                            },
                            "text/x-yaml": {
                                "schema": {
                                    "$ref": "#/components/schemas/v2_text_verse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "pagination": {
                "title": "Pagination",
                "description": "The new pagination meta response.",
                "properties": {
                    "pagination": {
                        "properties": {
                            "total": {
                                "type": "integer",
                                "example": 1801
                            },
                            "count": {
                                "type": "integer",
                                "example": 25
                            },
                            "per_page": {
                                "type": "integer",
                                "example": 25
                            },
                            "current_page": {
                                "type": "integer",
                                "example": 1
                            },
                            "total_pages": {
                                "type": "integer",
                                "example": 73
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "pagination"
                }
            },
            "v4_bible_timestamps": {
                "title": "Bible Timestamps",
                "description": "The bibles hash returned for timestamps",
                "type": "array",
                "items": {
                    "properties": {
                        "fileset_id": {
                            "$ref": "#/components/schemas/BibleFileset/properties/id"
                        }
                    },
                    "type": "object"
                },
                "xml": {
                    "name": "v4_bible.timestamps"
                }
            },
...
            "BibleFileset": {
                "title": "Bible Fileset",
                "description": "BibleFileset",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "The fileset id",
                        "type": "string",
                        "maxLength": 16,
                        "minLength": 6,
                        "example": "ENGESV"
                    },
                    "set_type_code": {
                        "$ref": "#/components/schemas/BibleFilesetType/properties/set_type_code"
                    },
                    "set_size_code": {
                        "$ref": "#/components/schemas/BibleFilesetSize/properties/set_size_code"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BibleFileset"
                }
            },
...
            "BibleLink": {
                "title": "BibleLink",
                "description": "BibleLink",
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "",
                        "type": "string"
                    },
                    "bible_id": {
                        "$ref": "#/components/schemas/Bible/properties/id"
                    },
                    "type": {
                        "title": "type",
                        "description": "The type of link, for example, web, pdf, print",
                        "type": "string",
                        "example": "web"
                    },
                    "url": {
                        "title": "url",
                        "description": "URL of linked content",
                        "type": "string",
                        "example": "http://bibles.org/versions/aai-AAINT"
                    },
                    "title": {
                        "title": "title",
                        "description": "",
                        "type": "string",
                        "example": "BibleSearch"
                    },
                    "organization_id": {
                        "$ref": "#/components/schemas/Organization/properties/id"
                    },
                    "created_at": {
                        "title": "created_at",
                        "description": "",
                        "type": "string"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "description": "",
                        "type": "string"
                    },
                    "provider": {
                        "title": "provider",
                        "description": "Organization providing the content referenced in the link",
                        "type": "string",
                        "example": "American Bible Society"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "BibleLink"
                }
            },
...
            "v4_bible.all": {
                "title": "v4_bible.all",
                "description": "The bibles being returned",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "abbr": {
                                    "$ref": "#/components/schemas/Bible/properties/id"
                                },
                                "name": {
                                    "$ref": "#/components/schemas/BibleTranslation/properties/name"
                                },
                                "vname": {
                                    "$ref": "#/components/schemas/BibleTranslation/properties/name"
                                },
                                "language": {
                                    "$ref": "#/components/schemas/Language/properties/name"
                                },
                                "language_autonym": {
                                    "$ref": "#/components/schemas/LanguageTranslation/properties/name"
                                },
                                "language_altNames": {
                                    "$ref": "#/components/schemas/LanguageTranslation/properties/name"
                                },
                                "iso": {
                                    "$ref": "#/components/schemas/Language/properties/iso"
                                },
                                "date": {
                                    "$ref": "#/components/schemas/Bible/properties/date"
                                },
                                "filesets": {
                                    "properties": {
                                        "dbp-prod": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/BibleFileset"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "meta": {
                        "$ref": "#/components/schemas/pagination"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "v4_bible.all"
                }
            },
...
            "v4_bible.one": {
                "title": "v4_bible.one",
                "description": "The bible being returned",
                "type": "array",
                "items": {
                    "properties": {
                        "abbr": {
                            "$ref": "#/components/schemas/Bible/properties/id"
                        },
                        "alphabet": {
                            "$ref": "#/components/schemas/Alphabet/properties/script"
                        },
                        "mark": {
                            "$ref": "#/components/schemas/Bible/properties/copyright"
                        },
                        "name": {
                            "$ref": "#/components/schemas/BibleTranslation/properties/name"
                        },
                        "description": {
                            "$ref": "#/components/schemas/BibleTranslation/properties/description"
                        },
                        "vname": {
                            "$ref": "#/components/schemas/BibleTranslation/properties/name"
                        },
                        "vdescription": {
                            "$ref": "#/components/schemas/BibleTranslation/properties/description"
                        },
                        "publishers": {
                            "$ref": "#/components/schemas/Organization"
                        },
                        "providers": {
                            "$ref": "#/components/schemas/Organization"
                        },
                        "language": {
                            "$ref": "#/components/schemas/Language/properties/name"
                        },
                        "iso": {
                            "$ref": "#/components/schemas/Language/properties/iso"
                        },
                        "date": {
                            "$ref": "#/components/schemas/Bible/properties/date"
                        },
                        "country": {
                            "$ref": "#/components/schemas/Country/properties/name"
                        },
                        "books": {
                            "$ref": "#/components/schemas/Book/properties/id"
                        },
                        "links": {
                            "$ref": "#/components/schemas/BibleLink"
                        },
                        "filesets": {
                            "$ref": "#/components/schemas/BibleFileset"
                        }
                    },
                    "type": "object"
                },
                "xml": {
                    "name": "v4_bible.one"
                }
            },
...
            "v2_text_verse": {
                "title": "v2_text_verse",
                "description": "The bible Search Group Response",
                "type": "array",
                "items": {
                    "properties": {
                        "book_name": {
                            "$ref": "#/components/schemas/Book/properties/name"
                        },
                        "book_id": {
                            "$ref": "#/components/schemas/Book/properties/name"
                        },
                        "book_order": {
                            "$ref": "#/components/schemas/Book/properties/protestant_order"
                        },
                        "chapter_id": {
                            "$ref": "#/components/schemas/BibleFile/properties/chapter_start"
                        },
                        "chapter_title": {
                            "type": "string",
                            "example": "Chapter 1"
                        },
                        "verse_id": {
                            "$ref": "#/components/schemas/BibleFile/properties/verse_start"
                        },
                        "verse_text": {
                            "$ref": "#/components/schemas/BibleFile/properties/verse_text"
                        },
                        "paragraph_number": {
                            "type": "string",
                            "example": "2"
                        }
                    },
                    "type": "object"
                },
                "xml": {
                    "name": "v2_text_verse"
                }
            },
...
        },
...
        "parameters": {
            "version_number": {
                "name": "v",
                "in": "query",
                "description": "The Version Number",
                "required": true,
                "schema": {
                    "type": "integer",
                    "enum": [
                        4,
                        2
                    ],
                    "example": 4
                }
            },
...
            "limit": {
                "name": "limit",
                "in": "query",
                "description": "The number of search results to return",
                "schema": {
                    "type": "integer",
                    "default": 25
                }
            },
            "page": {
                "name": "page",
                "in": "query",
                "description": "The current page of the results",
                "schema": {
                    "type": "integer",
                    "default": 1
                }
            },
...
        },
...
        "securitySchemes": {
            "dbp_key": {
                "type": "apiKey",
                "description": "The key granted to the API developer upon sign up",
                "name": "key",
                "in": "query"
            },
            "dbp_user_token": {
                "type": "apiKey",
                "description": "The token granted to an authenticated end user upon login",
                "name": "api_token",
                "in": "query"
            }
        }
    },
    "security": [
        {
            "dbp_key": []
        }
    ],
...
}